Added set commands, improved large bulk-data handling.
Added spop and scard set commands.

Added test_large_values test inspired by Anthony Lauzon.

rawDataReceived occasionally receives a chunk of bulk data that excludes
that last two bytes (\r\n). This is kind of suspicious, but in any case, it
is a symptom of this handler assuming it will get exactly the right amount
of raw data. It needs further improvement, namely it needs to actually
buffer the raw data.