Big change: An error response now errbacks the deferred.
Previously all ResponseError objects would always fire
the callback chain of the deferred instead of errbacking -
this is obviously not consistent and could lead to huge
problems. This has now been changed so that a Redis
error response will always errback the waiting deferred.

In addition the unit tests have been updated to reflect
this. In most cases the tests have been changed to explicity
use `failUnlessFailure`, and in some cases repetitive checks
have been removed.