Do NOT run unittests against local Redis on 6379.
Because the unittests contain steps to completely flush the default
db, these tests run the risk of accidentally deleting all data on a
production Redis instance. This is obviously not great. Instead,
the new default port the tests look for is 6381. If no redis server
is running on that port, all subsequent tests will be skipped and
an informative message will be printed explaining why these tests
were skipped. This way, errors won't be produced by the test suite
for this condition.

Also included are fixes for a few tests and a refactoring of the
'save' tests, which would sometimes fail if a BGSAVE was currently
in progress.