Support for variadic commands introduced in 2.4.
A variable number of values can now be passed to the following
commands: SADD, HDEL, SREM, ZREM, ZADD, L/RPUSH to bring support
in line with Redis 2.4.

A caveat - for ZADD the original implementation of txRedis broke
from convention of 'ZADD key score value' to have the method accept
'zadd(key, value, score)' (for whatever reason). This mode is still
supported for backwards compatibility but if a sequence of arguments
are passed, we now support a more Redis-like convention of
'zadd(key, score1, value1, score2, value2)'.